+++ src/tests/api/test_api_serializers.py
@@ -118,6 +118,7 @@ def test_submission_serializer_for_organiser(submission, orga_user):
         'track',
         'notes',
         'internal_notes',
+        'created',
     }
     assert isinstance(data['speakers'], list)
     assert data['speakers'][0] == {
@@ -128,6 +129,7 @@ def test_submission_serializer_for_organiser(submission, orga_user):
     }
     assert data['submission_type'] == str(submission.submission_type.name)
     assert data['slot'] is None
+    assert data['created'] == submission.created.astimezone(submission.event.tz).isoformat()
 
 
 @pytest.mark.django_db